Rename Stream.h back to StreamOptions.h
