+++ src/torchcodec/_core/StreamOptions.h
